+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[Unreleased]
+### Added
+- Save file size of the initial upload 
+
 [1.1.0] - 2018-04-24
 ### Changed
 - Namespace update to Level51\Cloudinary